// src/app/shared/models/ec-image/ec-image-model.ts
export class ECImageViewModel {
    ECImageId: number;
    OrigionalFilename: string;
    Filename: string;
    ImagePath: string;
    FilePath: string;
    ECImageTypeId: number;
    TypeName: string;
    UserId: string;
}


export class MinMax {
    Min: number;
    Max: number;
}

export class ECImageInfo {
    ImageUrl: string;
    FilePath: string;
    TypeName: string;
    Width: string;
    Height: string;
    HorizontalResolution: number;
    VerticalResolution: number;
}